initial additions
